2nd iteration will have no columns to work on
After for loop is done with 1st iteration, we need to add below lines for 2nd iteration to have some conditions as of 1st iteration.
For i = 1 To 2
 Cells.Select
 Selection.EntireColumn.Hidden = False
 If i = 2 Then Application.ScreenUpdating = False
NOTE: In absence of this, 2nd iterations will have no columns that will match if condition and thus its not the same situation. We need to unhide all columns before we start 2nd iteration.